backend/requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.2
